# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Quotation App
  version: 2.0.0
  description: Multi-tenant SaaS quotation & CRM platform for MSMEs
servers:
  - url: /api
...
security:
  - bearerAuth: []
paths:
...
  /quotations:
    get:
      tags: [Quotations]
      operationId: listQuotations
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: clientId
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Quotation" }
    post:
      tags: [Quotations]
      operationId: createQuotation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuotationInput" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quotation" }
...
  /leads:
    get:
      tags: [Leads]
      summary: List leads
      operationId: listLeads
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: priority
          schema: { type: string }
        - in: query
          name: source
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Lead" }
    post:
      tags: [Leads]
      summary: Create a lead
      operationId: createLead
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadInput" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lead" }
  /leads/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      tags: [Leads]
      operationId: getLead
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeadDetail" }
    patch:
      tags: [Leads]
      operationId: updateLead
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadInput" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lead" }
    delete:
      tags: [Leads]
      operationId: deleteLead
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
  /leads/{id}/activities:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      tags: [Leads]
      operationId: listLeadActivities
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LeadActivity" }
    post:
      tags: [Leads]
      operationId: createLeadActivity
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadActivityInput" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeadActivity" }
  /leads/{id}/score:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    post:
      tags: [Leads]
      summary: Re-score lead (AI + rules)
      operationId: scoreLead
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lead" }
  /leads/{id}/convert:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    post:
      tags: [Leads]
      summary: Convert lead to client (and optionally quotation)
      operationId: convertLead
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConvertLeadInput" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConvertLeadResponse" }

  # ── Tasks ──
...
  /payments:
    get:
      tags: [Payments]
      operationId: listPayments
      parameters:
        - in: query
          name: invoiceId
          schema: { type: integer }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Payment" }
    post:
      tags: [Payments]
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentInput" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
  /payments/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    delete:
      tags: [Payments]
      operationId: deletePayment
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }

  # ── Integrations ──
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }

    # ── Auth ──
...
    Quotation:
      type: object
      required:
        [id, quotationNumber, status, subtotal, discountAmount, taxAmount, total, createdAt, updatedAt]
      properties:
        id: { type: integer }
        quotationNumber: { type: string }
        clientId: { type: ["integer", "null"] }
        clientName: { type: ["string", "null"] }
        clientCompany: { type: ["string", "null"] }
        createdByName: { type: ["string", "null"] }
        status: { type: string, enum: [draft, sent, approved, rejected, expired] }
        validUntil: { type: ["string", "null"] }
        subtotal: { type: number }
        discountAmount: { type: number }
        discountPercent: { type: number }
        taxAmount: { type: number }
        taxPercent: { type: number }
        total: { type: number }
        notes: { type: ["string", "null"] }
        terms: { type: ["string", "null"] }
        itemCount: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    QuotationInput:
      type: object
      properties:
        clientId: { type: integer }
        validUntil: { type: string }
        notes: { type: string }
        terms: { type: string }
        discountPercent: { type: number }
        taxPercent: { type: number }
...
    Lead:
      type: object
      required: [id, name, source, status, priority, score, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        company: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        source: { type: string }
        externalId: { type: ["string", "null"] }
        status: { type: string }
        priority: { type: string }
        score: { type: integer }
        budget: { type: ["number", "null"] }
        product: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        nextAction: { type: ["string", "null"] }
        assignedToId: { type: ["integer", "null"] }
        convertedClientId: { type: ["integer", "null"] }
        lastContactedAt: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    LeadInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        company: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        source: { type: string, enum: [manual, indiamart, website, other] }
        status: { type: string, enum: [new, contacted, qualified, won, lost] }
        priority: { type: string, enum: [hot, warm, cold] }
        budget: { type: ["number", "null"] }
        product: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        nextAction: { type: ["string", "null"] }
        assignedToId: { type: ["integer", "null"] }
    LeadDetail:
      allOf:
        - $ref: "#/components/schemas/Lead"
        - type: object
          properties:
            activities:
              type: array
              items: { $ref: "#/components/schemas/LeadActivity" }
    LeadActivity:
      type: object
      required: [id, leadId, type, title, createdAt]
      properties:
        id: { type: integer }
        leadId: { type: integer }
        type: { type: string }
        title: { type: string }
        body: { type: ["string", "null"] }
        userId: { type: ["integer", "null"] }
        userName: { type: ["string", "null"] }
        createdAt: { type: string }
    LeadActivityInput:
      type: object
      required: [type, title]
      properties:
        type: { type: string, enum: [note, call, email, status_change, task, conversion] }
        title: { type: string }
        body: { type: ["string", "null"] }
    ConvertLeadInput:
      type: object
      properties:
        createQuotation: { type: boolean, default: false }
    ConvertLeadResponse:
      type: object
      required: [clientId]
      properties:
        clientId: { type: integer }
        quotationId: { type: ["integer", "null"] }
...
    Payment:
      type: object
      required: [id, invoiceId, amount, method, paidAt, createdAt]
      properties:
        id: { type: integer }
        invoiceId: { type: integer }
        amount: { type: number }
        method: { type: string }
        reference: { type: ["string", "null"] }
        paidAt: { type: string }
        notes: { type: ["string", "null"] }
        recordedByName: { type: ["string", "null"] }
        createdAt: { type: string }
    PaymentInput:
      type: object
      required: [invoiceId, amount]
      properties:
        invoiceId: { type: integer }
        amount: { type: number }
        method: { type: string, enum: [cash, upi, bank_transfer, cheque, card, other] }
        reference: { type: ["string", "null"] }
        paidAt: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
...