# lib/api-zod/src/generated/types/convertLeadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant SaaS quotation & CRM platform for MSMEs
 * OpenAPI spec version: 2.0.0
 */

export interface ConvertLeadInput {
  createQuotation?: boolean;
}