# lib/api-zod/src/generated/types/leadActivity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant SaaS quotation & CRM platform for MSMEs
 * OpenAPI spec version: 2.0.0
 */

export interface LeadActivity {
  id: number;
  leadId: number;
  type: string;
  title: string;
  /** @nullable */
  body?: string | null;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  userName?: string | null;
  createdAt: string;
}