# lib/api-zod/src/generated/types/leadActivityInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant SaaS quotation & CRM platform for MSMEs
 * OpenAPI spec version: 2.0.0
 */
import type { LeadActivityInputType } from './leadActivityInputType';

export interface LeadActivityInput {
  type: LeadActivityInputType;
  title: string;
  /** @nullable */
  body?: string | null;
}